Will use JSON for building parameters
The buildings can hopefully recide in the globals script. For example,
the CropFieldGlobals will contain the update information for all the
different crop fields since these should be similar to eachother. A json
loader whould go through each json file and create a object of this
class for each differen crop field